CPR: unnest classes and put into dedicated package
